+++ src/video/bink.cpp
@@ -636,7 +636,7 @@ void Bink::initAudioTrack(AudioTrack &audio) {
 
 	audio.overlapLen = audio.frameLen / 16;
 	audio.blockSize  = (audio.frameLen - audio.overlapLen) * audio.channels;
-	audio.root       = 2.0 / sqrt(audio.frameLen);
+	audio.root       = 2.0 / sqrt((float) audio.frameLen);
 
 	uint32 sampleRateHalf = (audio.sampleRate + 1) / 2;
